@@ -1,36 +1,22 @@
-import React from 'react';
-import { Routes, Route, Navigate } from 'react-router-dom';
-import Navbar from './components/navbar'; 
-import Home from './pages/home';
-import Properties from './pages/Properties';
-import PropertyDetails from './pages/propertyDetails';
-import { AuthProvider } from './context/authContext';
-import Bookings from './pages/bookings'; 
-import Login from './pages/login';
-import Register from './pages/register';
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import Navbar from "./assets/components/ui/navbar/navbar";
+import Home from "./assets/pages/home";
+import { AuthProvider } from "./context/auth-context";
+import "./App.css";
+import PropertyDetail from "./assets/pages/propertyDetail";
+import LoginForm from "./assets/pages/loginForm";
 
-function App() {
+export default function App() {
   return (
-    <AuthProvider>
-      <Navbar />
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/properties" element={<Properties />} />
-        <Route path="/properties/:id" element={<PropertyDetails />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/register" element={<Register />} />
-        
-        {/* Ruta principal de reservas */}
-        <Route path="/bookings" element={<Bookings />} />
-        
-        {/* Redirección desde singular a plural */}
-        <Route path="/booking" element={<Navigate to="/bookings" replace />} />
-        
-        {/* Captura todas las rutas no definidas */}
-        <Route path="*" element={<Navigate to="/" replace />} />
-      </Routes>
+    <AuthProvider> 
+      <BrowserRouter>
+        <Navbar />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/properties/:propertyId" element={<PropertyDetail />} />
+          <Route path="/login" element={<LoginForm />} />
+        </Routes>
+      </BrowserRouter>
     </AuthProvider>
   );
 }
-
-export default App;
