@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
-import { getProperties } from '../api';
+import { getProperties } from '../../services/api-service';
 
 function Home() {
   const [properties, setProperties] = useState([]);
@@ -22,11 +22,11 @@ function Home() {
   }, []);
 
   return (
-    <div className="min-h-screen bg-gray-100">
+    <div className="min-h-screen bg-gray-50">
       {/* Hero Section */}
-      <div className="bg-blue-600 text-white py-20">
-        <div className="max-w-4xl mx-auto text-center">
-          <h1 className="text-5xl font-bold mb-6">
+      <div className="bg-rose-500 text-white py-20">
+        <div className="max-w-4xl mx-auto text-center px-4">
+          <h1 className="text-5xl font-extrabold mb-6 tracking-tight">
             Encuentra tu alojamiento perfecto
           </h1>
           <p className="text-xl mb-8">
@@ -37,7 +37,7 @@ function Home() {
 
       {/* Featured Properties Section */}
       <div className="max-w-6xl mx-auto py-12">
-        <h2 className="text-3xl font-bold text-center mb-8">
+        <h2 className="text-3xl font-bold text-center text-gray-900 mb-8">
           Propiedades destacadas
         </h2>
         {isLoading ? (
@@ -47,21 +47,29 @@ function Home() {
         ) : (
           <div className="grid grid-cols-1 md:grid-cols-3 gap-6 px-4">
             {properties.map((property) => (
-              <div key={property._id} className="bg-white rounded-lg shadow-lg overflow-hidden hover:scale-105 transition-transform duration-300">
+              <div
+                key={property._id}
+                className="bg-white rounded-lg shadow-lg overflow-hidden transform hover:scale-105 transition duration-300"
+              >
                 <img
                   src={property.photo || '/default-property.jpg'}
                   alt={property.title}
                   className="w-full h-48 object-cover"
                 />
-                <div className="p-6">
-                  <h3 className="text-xl font-semibold mb-2">{property.title}</h3>
-                  <p className="text-gray-600 mb-4">{property.location}</p>
-                  <Link
-                    to={`/properties/${property._id}`}
-                    className="text-blue-600 hover:text-blue-500 font-semibold"
-                  >
-                    Ver detalles
-                  </Link>
+                <div className="p-6 flex flex-col gap-2">
+                  <h3 className="text-xl font-semibold text-gray-900">{property.title}</h3>
+                  <p className="text-gray-600">{property.location}</p>
+                  <div className="flex items-center justify-between mt-2">
+                    <span className="text-lg font-bold text-rose-500">
+                      {property.price}€ / noche
+                    </span>
+                    <Link
+                      to={`/properties/${property._id}`}
+                      className="px-4 py-2 bg-rose-500 text-white rounded-lg hover:bg-rose-600 transition"
+                    >
+                      Reservar
+                    </Link>
+                  </div>
                 </div>
               </div>
             ))}
